refactor(cognition): extract perception predicates from act_observe into their own module
Pure code-motion, behavior-identical. The five structural-fact DETECTORS
(collect_touched_paths, entries_since_last_settlement, all_calls_already_satisfied,
is_orientation_call, is_redundant_orientation) — pure predicates over the working-
memory trail + a tool batch — move to act_observe/perception.rs (127 lines). They are
one concern (proprioception, not policy) and now read as one file.

act_observe/mod.rs: 3,011 -> 2,899. Also removed an orphan doc comment left behind by
the recency extraction (boy-scout).

cargo check green; 28/28 act_observe module tests pass unchanged. The discipline is the
guardrail: pure code-motion + green at every step means a bad cut shows as a red test on
one file, never a silent runtime break.
